+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - Bump constraint on python to >=3.8,<3.12 (upper constraint is required by pex)
 - Update dependencies
 - [jumpcloud,okta] Verify SSO URLs before making SAML requests
+- [jumpcloud,okta] Verify that plugin is configured before making SAML requests
 
 ## 0.4.0 - 2022-08-02